Create Locale.ru.resx
Russian Localization:
Contains errors - some verbs are written with a capital letter. In Russian and Ukrainian, this is considered an error. In English in the original, was capitalized, and I also left it temporarily with a capital letter. We need to see how, it will be in the interface of the application.